Adjust tests for coverage
This change tweaks the tests to ensure we are capturing the newly added
try/catch at the top of the `.run()` method.
